services/market-data/requirements.txt
ccxt>=4.2.0
aiohttp>=3.9.0
asyncio-mqtt>=0.13.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.5.0
loguru>=0.7.0